+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - add required configuration, initial files and folders
 - add classes and methods to request tankerkönig api and store values in sqlite database
+- add internal api to retrieve the price data from database as JSON
 
 ### Changed